HII-17: types abstraction.
